Cast open and close levels to float

+++ config/EURUSD_H1.h
@@ -10,10 +10,10 @@ struct Stg_AC_Params_H1 : StgParams {
     lot_size = 0;
     signal_open_method = 0;
     signal_open_filter = 1;
-    signal_open_level = 2;
+    signal_open_level = (float)2;
     signal_open_boost = 0;
     signal_close_method = 0;
-    signal_close_level = 2;
+    signal_close_level = (float)2;
     price_stop_method = 0;
     price_stop_level = 2;
     tick_filter_method = 1;
